Fixed an integer type for packet streaming

+++ network/sfml/packets.hpp
@@ -16,17 +16,17 @@ sf::Packet & operator>>(sf::Packet & p, gui::GameState::Snake & snake);
 // Vectors
 template <class T>
 sf::Packet & operator>>(sf::Packet & p, std::vector<T> & v) {
-    sf::Uint64 size;
+    sf::Uint32 size;
     p >> size;
     v.resize(size);
-    for (sf::Uint64 i = 0; i < size; ++i)
+    for (sf::Uint32 i = 0; i < size; ++i)
         p >> v[i];
     return p;
 }
 
 template <class T>
 sf::Packet & operator<<(sf::Packet & p, const std::vector<T> & v) {
-    p << static_cast<sf::Uint64>(v.size());
+    p << static_cast<sf::Uint32>(v.size());
     for (const auto & t: v)
         p << t;
     return p;
